reorganize code v1
